Moves errors to separate module
